+++ webgoat-lessons/password-reset/src/main/resources/html/PasswordReset.html
@@ -140,6 +140,37 @@
 </div>
 
 <div class="lesson-page-wrapper">
+    <div class="adoc-content" th:replace="doc:PasswordReset_SecurityQuestions.adoc"></div>
+    <div class="attack-container">
+        <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
+        <form class="attack-form" accept-charset="UNKNOWN"
+              method="POST" name="form"
+              action="/WebGoat/PasswordReset/SecurityQuestions"
+              enctype="application/json;charset=UTF-8">
+            <select name="question">
+                <option>What is your favorite animal?</option>
+                <option>In what year was your mother born?</option>
+                <option>What was the time you were born?</option>
+                <option>What is the name of the person you first kissed?</option>
+                <option>What was the house number and street name you lived in as a child?</option>
+                <option>In what town or city was your first full time job?</option>
+                <option>In what city were you born?</option>
+                <option>On which wrist do you were your watch?</option>
+                <option>What was the last name of your favorite teacher in grade three?</option>
+                <option>What is the name of a college/job you applied to but didn't attend?</option>
+                <option>What are the last 5 digits of your drivers license?</option>
+                <option>What was your childhood nickname?</option>
+                <option>Who was your childhood hero?</option>
+                <option>What is your favorite color?</option>
+            </select>
+            <input name="Check Question" value="check" type="SUBMIT"/>
+        </form>
+        <br/>
+        <div class="attack-output"></div>
+    </div>
+</div>
+
+<div class="lesson-page-wrapper">
     <div class="adoc-content" th:replace="doc:PasswordReset_host_header.adoc"></div>
     <div class="attack-container">
         <img th:src="@{/images/wolf-enabled.png}" class="webwolf-enabled"/>
@@ -234,37 +265,5 @@
 
 <div class="lesson-page-wrapper">
     <div class="adoc-content" th:replace="doc:PasswordReset_mitigation.adoc"></div>
-</div>
-
-<div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:PasswordReset_SecurityQuestions.adoc"></div>
-    <div class="attack-container">
-        <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
-        <form class="attack-form" accept-charset="UNKNOWN"
-              method="POST" name="form"
-              action="/WebGoat/PasswordReset/SecurityQuestions"
-              enctype="application/json;charset=UTF-8">
-            <select name="question">
-                <option>What is your favorite animal?</option>
-                <option>In what year was your mother born?</option>
-                <option>What was the time you were born?</option>
-                <option>What is the name of the person you first kissed?</option>
-                <option>What was the house number and street name you lived in as a child?</option>
-                <option>In what town or city was your first full time job?</option>
-                <option>In what city were you born?</option>
-                <option>On which wrist do you were your watch?</option>
-                <option>What was the last name of your favorite teacher in grade three?</option>
-                <option>What is the name of a college/job you applied to but didn't attend?</option>
-                <option>What are the last 5 digits of your drivers license?</option>
-                <option>What was your childhood nickname?</option>
-                <option>Who was your childhood hero?</option>
-                <option>What is your favorite color?</option>
-            </select>
-            <input name="Check Question" value="check" type="SUBMIT"/>
-        </form>
-        <br/>
-        <div class="attack-feedback"></div>
-        <div class="attack-output"></div>
-    </div>
 </div>
 </html>
