+++ webgoat-lessons/password-reset/src/main/resources/lessonPlans/en/PasswordReset_SecurityQuestions.adoc
@@ -1,17 +1,16 @@
-== Choosing a Security Question
+== The Problem with Security Questions
 
-We have already talked about Security questions a bit. A good security question should meet the following criteria:
+While Security Questions my at first seem like a good way for authentication of a user, they
+have some big problems.
 
-- Safe: The answer should not be easy to research or guess.
-- Stable: The answer should be stable, meaning that it is not subject to change.
-- Memorable: The answer should be easy to remember.
-- Simple: The question should be: precise, easy and consistent.
-- Many: The question should have many possible answers.
+The "perfect" Security Question should be hard to crack, but easy to remember. Also the answer needs to fixed,
+so the answer must not be subject to change.
 
-== Try It! Choosing a  good security question.
+There are only a handful of questions which satisfy these criteria and practically none which apply to anybody.
 
-In this assignment your goal is to good security question from the dropdown list below.
-The Assignment is complete when you picked a security question which is considered good.
+If you have to pick a security question, we recommend not answering them truthfully.
 
-Note: Some may say that one question is better than another, so this list is a bit subjective.
-      But you should not be having any problem differencing between the good and bad.
+To further elaborate on the matter, there is a small assignment for you: There is a list of some common security questions.
+if you choose one, it will show to you why the question you picked is not really as good as one may think.
+
+When you have looked at two questions the assignment will be marked as complete.
